+++ ghi/app/src/Features/MainPage/MainPage.js
@@ -1,4 +1,4 @@
-import Carousel from "./carousel";
+import Carousel from "../Carousel/carousel";
 
 function MainPage() {
   return (
